Desktop/ResumeMatcher/backend/requirements.txt
flask
flask-cors
scikit-learn
pdfplumber
python-docx

